+++ src/main/java/net/resultadofinal/micomercial/model/Ingrediente.java
@@ -1,9 +1,11 @@
 package net.resultadofinal.micomercial.model;
 
+import java.math.BigDecimal;
+
 public class Ingrediente {
     private Short id;
     private Long productoId, ingredienteId;
-    private Integer cantidad;
+    private BigDecimal cantidad;
     private String xproducto, xtipo, xmedida;
 
     public Long getIngredienteId() {
@@ -30,11 +32,11 @@ public class Ingrediente {
         this.productoId = productoId;
     }
 
-    public Integer getCantidad() {
+    public BigDecimal getCantidad() {
         return cantidad;
     }
 
-    public void setCantidad(Integer cantidad) {
+    public void setCantidad(BigDecimal cantidad) {
         this.cantidad = cantidad;
     }
